Exceptions handling

+++ src/main/java/Backend/Caracteristicas/CaracteristicaServiceImpl.java
@@ -1,20 +1,27 @@
 package Backend.Caracteristicas;
 
+import Backend.exceptions.CaracteristicaNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+
 @Service
 public class CaracteristicaServiceImpl implements CaracteristicaService {
 
     @Autowired
     private CaracteristicaRepository caracteristicaRepository;
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(CaracteristicaServiceImpl.class);
+
     @Override
     public void CrearCaracteristica(Caracteristica caracteristica) {
         caracteristicaRepository.save(caracteristica);
-
     }
 
     @Override
@@ -30,11 +37,20 @@ public class CaracteristicaServiceImpl implements CaracteristicaService {
     @Override
     public void ModificarCaracteristica(Caracteristica caracteristica) {
         caracteristicaRepository.save(caracteristica);
-
     }
 
     @Override
     public void EliminarCaracteristica(Long id) {
-        caracteristicaRepository.deleteById(id);
+        Optional<Caracteristica> optionalCaracteristica = caracteristicaRepository.findById(id);
+        if (optionalCaracteristica.isPresent()) {
+            try {
+                caracteristicaRepository.deleteById(id);
+            } catch (Exception e) {
+                LOGGER.error("Error eliminando característica con id: {}", id, e);
+                throw new RuntimeException("Error eliminando característica con id: " + id, e);
+            }
+        } else {
+            throw new CaracteristicaNotFoundException("No se encontró una característica con id: " + id);
+        }
     }
 }
